Teleport offset fix (#1404)
Fix offset behavior to set the player's z offset before clamping with boom teleport specials (207-210, 268 and 269)

Add tests for functionality including previous bugged behavior that would stick the player above the ceiling

Fix WorldAllocator not validating the extracted wad file exists

+++ Tests/Unit/GameAction/WorldAllocator.cs
@@ -59,6 +59,9 @@ internal static class WorldAllocator
         var profiler = new Profiler();
         var audioSystem = new MockAudioSystem();
 
+        if (!File.Exists(fileName))
+            throw new Exception($"The file {fileName} does not exist");
+
         string iwadFileName = IWadInfo.GetDefaultFileName(iwadType);
         if (!File.Exists(iwadFileName))
             File.Copy("Resources/dummy.wad", iwadFileName, true);
